mentor/pyproject.toml
[project]
name = "mentor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]

[project.scripts]
mentor = "mentor:main"
ai-mentor = "mentor.my_mentor:mentor_agent"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
